ffs: fix path parsing when the root is base64
Since base64 may include a slash "/" which is also used for path separation,
check first for the presence of an "=". If one occurs at the end of the spec or
just before the first slash, treat the stuff to its left as the root instead of
splitting inside it.

This is still not perfect; a spec may have an "=" elsewhere too. A better
solution would be to use a different delimiter. But slash is convenient, and I
think that case is not super likely, so let's see how it goes.