ffstools: use separate keyspaces for roots and non-root blobs (#11)
Use prefixed store wrappers to partition the keys for roots from
the keys for other blobs. Roots now begin with "@" and other keys
begin with " " (space).

This generalizes the existing convention that keys beginning with
the string "root:" were supposed to be root blobs. That works fine
as long as we are lucky, but is not very robust.

Related changes:

- Add a -bucket global flag to the blob tool.
- Update the ffs module to import prefixed package improvements.
- Update the ffs config package to support prefixes.
- Update ffs subcommands to be aware of the prefixed storage.