ffs: add support for extended attributes to "file ls"
Add a "@" marker like macOS uses to indicate the presence of extended
attributes, and an -xattr flag to print their names and lengths in the textual
listing.

In JSON output with -xattr set, include the extended attributes as an object.
As usual, values are base64-encoded to support binary values.