ffs: avoid repeated work when indexing
Don't re-scan files we have already scanned. We keep track of this separately
from the index because there is a small chance the key could be a false
positive in the Bloom filter, and we don't want to lose keys during garbage
collection. The scanned check is ephemeral and not persisted.