ffs: add -partial flag to the gc subcommand
Sweeping a large cloud store can take a long time, because the API calls to
delete stuff are often very slow. The -partial flag allows sweeping a subset of
the eligible objects, which improves the situation without taking so long. As
long as you pick a fraction bigger than the steady state growth rate, it will
converge toward a good state.