ffs: make the sync command smarter about scanning remote keys
If the remote store has a large keyspace and the worklist is small, there is no
point fetching the whole key list just to filter it. So, if the worklist is
below a cutoff (currently 500) don't even try to filter it, just push them all.

Above the cutoff, be smarter about which ranges of the key space to scan: For
each 1-byte prefix, find the minimum and maximum key in the worklist with that
prefix and only scan within those specific ranges. In the limit we will still
scan most of the key space, but in many cases we can get away with much less.