blobd: remove the tailscale dependency
Maintaining this is too much of a pain, the tailscale.com module is very
sensitive to subtle version differences and breaks when packages that it cares
about are updated separately from its own pinned versions. This is particularly
bad for gvisor, but happens for some of the golang packages too.

Rather than bake tsnet directly into blobd, a user can use "tailscale serve" to
expose a plain blobd listener to the tailnet. That's just as easy to set up,
and means I don't have to depend on user-hostile nonsense like gvisor and otel.