all: update ffs to v0.9.1 and use the new API (#37)
all: update ffs to v0.9.1 and fix all affected packages

Most of the functionality does not change, with a few exceptions:

1. The blob cas-put and cas-key subcommands are removed.
   This whole subcommand should probably be reworked, but later.

2. The storage server is no longer compatible with existing stores.
   Any data saved in the previous affixed format must be migrated.

Otherwise, most of the changes are internal to the plumbing.

The config.CAS type was renamed config.Store.
This change has to be done in two parts, because the ffuse driver depends on
config.CAS and cannot be updated till the config package is updated. So stick
in a type alias to make the build work, and remove it once ffuse is updated.