lib/storeservice: make it possible to assign different stores per caller
Instead of giving the same root store to each caller, introduce a check that
can choose a different store based on who connects. For now it always reports
the same root store (as before). In future commits, we can wire this up to
check Tailscale identity and assign a different substore based on that.