ffs: speed up and reduce the size of indexes
Previously, we would estimate the maximum size of the index using the total
number of keys in the bucket. That forces the index to contain in general many
more segments than it actually needs for the number of keys in the root being
indexed.

Instead, precompute the exact set of keys that need to go into the index,
allowing the index to be a lot smaller. This also has the nice side effect of
removing redundant adds of data keys, which speeds up the index generation.