ffs/cmdstorage: add keyring support
Change getEncryptionKey to handle keyring.Ring as well as the existing plain
file and keyfile.File support.

Treat the existing key formats as single-key Rings for purposes of talking to
the storeservice package. This preserves the existing behaviour, since all the
blobs written prior to the change have no key ID.